working offscreen rendering
